requirements.txt
numpy==1.19.4
torch==1.6.0
torchvision==0.7.0
tensorboard==2.3.0
tqdm==4.51.0
pillow==7.2.0 
matplotlib==3.3.1